done with reorder list
