+++ lib/Redesign/Testin_components/building/MyReorder.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'reorder2.dart';
 
 class B2 extends StatelessWidget {
   const B2({super.key});
@@ -9,11 +8,6 @@ class B2 extends StatelessWidget {
     return const Scaffold(
       body: Row(
         children: [
-          CreateRecipeListCard(),
-          SizedBox(
-            width: 600,
-            child: ReorderIngridientlist(),
-          ),
           CreateRecipeListCard2(),
         ],
       ),
@@ -21,23 +15,23 @@ class B2 extends StatelessWidget {
   }
 }
 
-class CreateRecipeListCard extends StatelessWidget {
-  const CreateRecipeListCard({
+class CreateRecipeListCard2 extends StatelessWidget {
+  const CreateRecipeListCard2({
     super.key,
   });
 
   @override
   Widget build(BuildContext context) {
     return SizedBox(
-      width: 410,
-      height: 463,
+      width: 500,
+      height: 450,
       child: Card(
         color: const Color(0xffd9d9d9),
         child: ClipRRect(
           borderRadius: BorderRadius.circular(10),
-          child: Column(
+          child: const Column(
             children: [
-              const CreateIngMainTile(),
+              CreateIngMainTile(),
               Expanded(
                 child: ReorderIngridientlist(),
               ),
@@ -49,38 +43,6 @@ class CreateRecipeListCard extends StatelessWidget {
   }
 }
 
-class CreateIngMainTile extends StatelessWidget {
-  const CreateIngMainTile({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    const textStyle = TextStyle(
-      fontSize: 24,
-      fontWeight: FontWeight.w700,
-      color: Color(0xFF000000), // Set your desired color here
-    );
-
-    return Container(
-      height: 50,
-      width: 410,
-      color: const Color(0xFFD9D9D9),
-      child: Row(
-        children: [
-          const SizedBox(width: 20),
-          Image.asset('images/new/icons/cart.png'),
-          const SizedBox(width: 82),
-          const Text("Ingridients", style: textStyle),
-          const SizedBox(width: 82),
-          Image.asset('images/new/icons/cart.png'),
-        ],
-      ),
-    );
-  }
-}
-
-//
 class ReorderIngridientlist extends StatefulWidget {
   const ReorderIngridientlist({super.key});
 
@@ -90,8 +52,8 @@ class ReorderIngridientlist extends StatefulWidget {
 
 class _ReorderIngridientlistState extends State<ReorderIngridientlist> {
   final List<Item> _items = [
-    Item(label: 'Item 1', amount: '100', unit: 'kg'),
-    Item(label: 'Item 2', amount: '200', unit: 'g'),
+    Item(label: 'maricha sås', amount: '1000', unit: 'ml'),
+    Item(label: 'Jordgubbssylt', amount: '2000', unit: 'tsk'),
     // ... more items
   ];
   @override
@@ -100,7 +62,7 @@ class _ReorderIngridientlistState extends State<ReorderIngridientlist> {
       buildDefaultDragHandles: false,
       children: List.generate(_items.length, (index) {
         Item item = _items[index];
-        return EditingTile(
+        return Tilebbs(
           key: ValueKey(item.label),
           label: item.label,
           amount: item.amount,
@@ -130,7 +92,62 @@ class Item {
   Item({required this.label, required this.amount, required this.unit});
 }
 
-class EditingTile extends StatelessWidget {
+class CreateIngMainTile extends StatelessWidget {
+  const CreateIngMainTile({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    const textStyle = TextStyle(
+      fontSize: 24,
+      fontWeight: FontWeight.w700,
+      color: Color(0xFF000000), // Set your desired color here
+    );
+
+    return Container(
+      height: 50,
+      width: 500,
+      color: const Color(0xFFD9D9D9),
+      child: const Stack(
+        children: [
+          Positioned(
+            left: 20,
+            top: 10,
+            child: carticon(),
+          ),
+          Positioned(
+            top: 10,
+            left: 0,
+            right: 0,
+            child: Align(
+              alignment: Alignment.center,
+              child: Text("Ingridients", style: textStyle),
+            ),
+          ),
+          Positioned(
+            top: 10,
+            right: 20,
+            child: carticon(),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class carticon extends StatelessWidget {
+  const carticon({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Image.asset('images/new/icons/cart.png');
+  }
+}
+
+class Tilebbs extends StatelessWidget {
   final String label; // First parameter
   final String amount;
   final String unit;
@@ -139,7 +156,7 @@ class EditingTile extends StatelessWidget {
   // Third parameter
 
   // Update the constructor to accept three parameters
-  const EditingTile({
+  const Tilebbs({
     super.key,
     // Initialize key
     required this.label,
@@ -158,49 +175,90 @@ class EditingTile extends StatelessWidget {
 
     return Container(
       height: 50,
-      width: 410,
+      width: 500,
       color: const Color(0xff161414),
       child: SizedBox(
-        child: Row(
+        child: Stack(
           children: [
-            const SizedBox(
-              width: 16,
-            ),
-            ReorderableDragStartListener(
-              index: index,
-              child: IconButton(
-                onPressed: () {
-                  print('menue clicked');
-                },
-                icon: const Icon(
-                  Icons.menu,
-                  color: Color.fromARGB(255, 233, 228, 228),
+            Positioned(
+              top: 5,
+              left: 5,
+              child: ReorderableDragStartListener(
+                index: index,
+                child: IconButton(
+                  onPressed: () {
+                    print('menue clicked');
+                  },
+                  icon: const Icon(
+                    Icons.menu,
+                    color: Color.fromARGB(255, 233, 228, 228),
+                  ),
                 ),
               ),
             ),
-            const SizedBox(width: 16),
-            Text(label, style: textStyle), // Use the label parameter
-            const Text(':', style: textStyle),
-            const SizedBox(width: 20),
-            Text(amount, style: textStyle), // Use the quantity parameter
-            const SizedBox(width: 3),
-            Text(unit, style: textStyle), // Use the unit parameter
-            const SizedBox(
-              width: 120,
-            ),
-            IconButton(
-              constraints: const BoxConstraints(maxHeight: 100, maxWidth: 100),
-              icon: const Icon(Icons.edit, size: 24),
-              color: const Color.fromARGB(255, 233, 228, 228),
-              // Plus icon
-              onPressed: () {
-                print('Edit clicked');
-                // Add your action for this button
-              },
+            Positioned(
+                top: 5,
+                left: 45,
+                child: Text('$label:',
+                    style: textStyle)), // Use the label parameter
+            Positioned(
+                top: 5,
+                left: 300,
+                child: Text('$amount $unit',
+                    style: textStyle)), // Use the quantity parameter
+
+            // Use the unit parameter
+
+            const Positioned(
+              top: 5,
+              right: 5,
+              child: Row(
+                children: [Editcontainer(), Deletecontainer()],
+              ),
             ),
           ],
         ),
       ),
+    );
+  }
+}
+
+class Deletecontainer extends StatelessWidget {
+  const Deletecontainer({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return IconButton(
+      constraints: const BoxConstraints(maxHeight: 100, maxWidth: 100),
+      icon: const Icon(Icons.delete, size: 24),
+      color: const Color.fromARGB(255, 233, 228, 228),
+      // Plus icon
+      onPressed: () {
+        print('delete button');
+        // Add your action for this button
+      },
+    );
+  }
+}
+
+class Editcontainer extends StatelessWidget {
+  const Editcontainer({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return IconButton(
+      constraints: const BoxConstraints(maxHeight: 100, maxWidth: 100),
+      icon: const Icon(Icons.edit, size: 24),
+      color: const Color.fromARGB(255, 233, 228, 228),
+      // Plus icon
+      onPressed: () {
+        print('Edit clicked');
+        // Add your action for this button
+      },
     );
   }
 }
